refactor(react): update radio tile tests to rtl
